Master

See merge request sudeepa-nisal/event-management-system!14

+++ src/main/java/org/example/event_management/service/EventService.java
@@ -34,8 +34,12 @@ public class EventService {
             return eventRepository.save(event);
         });
     }
-    @Transactional(readOnly = true)
-    public void deleteEvent(Long id) {
-        eventRepository.deleteById(id);
+//    @Transactional(readOnly = true)
+    public boolean deleteEvent(Long id) {
+        if (eventRepository.existsById(id)) {
+            eventRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
